+++ tests/reference/PostgreSQL_Builder_create_Hierarchy_connected_10.10.0.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS "ns_unittests"."Hierarchy" (
-  "managerId" integer NOT NULL, 
-  "manageeId" integer NOT NULL, 
+  "managerId" bigint NOT NULL, 
+  "manageeId" bigint NOT NULL, 
   PRIMARY KEY ("managerId", "manageeId"), 
   FOREIGN KEY ("managerId") REFERENCES "ns_unittests"."Employees" ("id") ON UPDATE CASCADE ON DELETE CASCADE, 
   FOREIGN KEY ("manageeId") REFERENCES "ns_unittests"."Employees" ("id") ON UPDATE CASCADE ON DELETE CASCADE
